requirements.txt
simpleeval
django
djangorestframework